move combat to util
