+++ src/main/java/sypztep/dominatus/mixin/core/item/refinement/newdurability/ItemStackMixin.java
@@ -6,7 +6,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import sypztep.dominatus.common.data.DominatusItemEntry;
-import sypztep.dominatus.common.util.RefineSystem.RefinementManager;
+import sypztep.dominatus.common.util.refinesystem.RefinementManager;
 
 @Mixin(ItemStack.class)
 public class ItemStackMixin {
